fix(mssql): raw return type
