+++ lib/dialects/mssql/query.js
@@ -153,21 +153,8 @@ class Query extends AbstractQuery {
     let result = this.instance;
     if (this.isInsertQuery(data)) {
       this.handleInsertQuery(data);
-
-      if (!this.instance) {
-        if (this.options.plain) {
-          // NOTE: super contrived. This just passes the newly added query-interface
-          //       test returning only the PK. There isn't a way in MSSQL to identify
-          //       that a given return value is the PK, and we have no schema information
-          //       because there was no calling Model.
-          const record = data[0];
-          result = record[Object.keys(record)[0]];
-        } else {
-          result = data;
-        }
-      }
+      return [result, rowCount];
     }
-
     if (this.isShowTablesQuery()) {
       return this.handleShowTablesQuery(data);
     }
@@ -197,7 +184,6 @@ class Query extends AbstractQuery {
             result[_result.Name].type += `(${_result.Length})`;
           }
         }
-
       }
     }
     if (this.isSelectQuery()) {
@@ -224,17 +210,15 @@ class Query extends AbstractQuery {
     if (this.isUpsertQuery()) {
       return [result, data[0].$action === 'INSERT'];
     }
-    if (this.isInsertQuery() || this.isUpdateQuery()) {
+    if (this.isUpdateQuery()) {
       return [result, rowCount];
     }
     if (this.isShowConstraintsQuery()) {
       return this.handleShowConstraintsQuery(data);
     }
     if (this.isRawQuery()) {
-      // MSSQL returns row data and metadata (affected rows etc) in a single object - let's standarize it, sorta
-      return [data, data];
+      return [data, rowCount];
     }
-
     return result;
   }
